tests/wpt/web-platform-tests/dom/nodes/Document-URL.sub.html

<!doctype html>
<meta charset=utf-8>
<title>Document.URL with redirect</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var iframe = document.createElement("iframe");
  iframe.src = "/common/redirect.py?location=/common/blank.html";
  document.body.appendChild(iframe);
  this.add_cleanup(function() { document.body.removeChild(iframe); });
  iframe.onload = this.step_func_done(function() {
    assert_equals(iframe.contentDocument.URL,
                  "http://{{host}}:{{ports[http][0]}}/common/blank.html");
  });
})
</script>
